apps/backend/pyproject.toml
[tool.poetry]
name = "docu-watch-backend"
version = "0.1.0"
description = "Backend service for DocuWatch application"
authors = ["kiranbaby14 <kiranbaby256@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.6"}
docusign-esign = "^4.0.0"
itsdangerous = "^2.2.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pydantic-settings = "^2.7.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
